Use arc4random() and friends instead of random() if on OpenBSD
Quells warnings when building about random() being bad.

It may be the case that other platforms have arc4random() too, but
I don't know which they are, and I know at least one does not, so
to keep portability the changes have been ifdef'd to OpenBSD only.

Adding arc4random() to openbsd-compat is non-trivial, because it
uses the kern.random sysctl when stirring, which is not present
in most systems, and relying on other sources (e.g. /dev/random)
have been known to cause issues not obvious to the user.